@@ -2,6 +2,7 @@
   "name": "opendtu",
   "version": "0.1.0",
   "private": true,
+  "type": "module",
   "scripts": {
     "dev": "vite",
     "build": "run-p type-check build-only",
@@ -17,30 +18,30 @@
     "mitt": "^3.0.1",
     "sortablejs": "^1.15.0",
     "spark-md5": "^3.0.2",
-    "vue": "^3.3.4",
-    "vue-i18n": "^9.5.0",
+    "vue": "^3.3.8",
+    "vue-i18n": "^9.7.0",
     "vue-router": "^4.2.5"
   },
   "devDependencies": {
-    "@intlify/unplugin-vue-i18n": "^1.4.0",
+    "@intlify/unplugin-vue-i18n": "^1.5.0",
     "@rushstack/eslint-patch": "^1.5.1",
     "@tsconfig/node18": "^18.2.2",
-    "@types/bootstrap": "^5.2.7",
-    "@types/node": "^20.8.3",
-    "@types/sortablejs": "^1.15.3",
-    "@types/spark-md5": "^3.0.2",
-    "@vitejs/plugin-vue": "^4.4.0",
+    "@types/bootstrap": "^5.2.9",
+    "@types/node": "^20.9.0",
+    "@types/sortablejs": "^1.15.5",
+    "@types/spark-md5": "^3.0.4",
+    "@vitejs/plugin-vue": "^4.5.0",
     "@vue/eslint-config-typescript": "^12.0.0",
     "@vue/tsconfig": "^0.4.0",
-    "eslint": "^8.51.0",
-    "eslint-plugin-vue": "^9.17.0",
+    "eslint": "^8.53.0",
+    "eslint-plugin-vue": "^9.18.1",
     "npm-run-all": "^4.1.5",
-    "sass": "^1.69.0",
-    "terser": "^5.21.0",
+    "sass": "^1.69.5",
+    "terser": "^5.24.0",
     "typescript": "^5.2.2",
-    "vite": "^4.4.11",
+    "vite": "^5.0.0",
     "vite-plugin-compression": "^0.5.1",
     "vite-plugin-css-injected-by-js": "^3.3.0",
-    "vue-tsc": "^1.8.18"
+    "vue-tsc": "^1.8.22"
   }
 }
