@@ -1,9 +1,8 @@
 // SPDX-License-Identifier: GPL-2.0-or-later
 #pragma once
 
+#include <mutex>
 #include <sunset.h>
-
-#define SUNPOS_UPDATE_INTERVAL 60000l
 
 class SunPositionClass {
 public:
@@ -15,18 +14,22 @@ public:
     bool isSunsetAvailable();
     bool sunsetTime(struct tm* info);
     bool sunriseTime(struct tm* info);
+    void setDoRecalc(bool doRecalc);
 
 private:
     void updateSunData();
+    bool checkRecalcDayChanged();
+    bool getDoRecalc();
 
     SunSet _sun;
-    bool _isDayPeriod = true;
     bool _isSunsetAvailable = true;
     uint32_t _sunriseMinutes = 0;
     uint32_t _sunsetMinutes = 0;
 
-    uint32_t _lastUpdate = 0;
     bool _isValidInfo = false;
+    bool _doRecalc = true;
+    std::mutex _recalcLock;
+    uint32_t _lastSunPositionCalculatedYMD = 0;
 };
 
 extern SunPositionClass SunPosition;
