@@ -8,7 +8,7 @@
 #define ALARM_LOG_ENTRY_SIZE 12
 #define ALARM_LOG_PAYLOAD_SIZE (ALARM_LOG_ENTRY_COUNT * ALARM_LOG_ENTRY_SIZE + 4)
 
-#define ALARM_MSG_COUNT 111
+#define ALARM_MSG_COUNT 130
 
 struct AlarmLogEntry_t {
     uint16_t MessageId;
@@ -22,10 +22,18 @@ enum class AlarmMessageType_t {
     HMT
 };
 
+enum class AlarmMessageLocale_t {
+    EN,
+    DE,
+    FR
+};
+
 typedef struct {
     AlarmMessageType_t InverterType;
     uint16_t MessageId;
-    String Message;
+    char Message_en[62];
+    char Message_de[63];
+    char Message_fr[64];
 } AlarmMessage_t;
 
 class AlarmLogParser : public Parser {
@@ -35,7 +43,7 @@ public:
     void appendFragment(uint8_t offset, uint8_t* payload, uint8_t len);
 
     uint8_t getEntryCount();
-    void getLogEntry(uint8_t entryId, AlarmLogEntry_t* entry);
+    void getLogEntry(uint8_t entryId, AlarmLogEntry_t* entry, AlarmMessageLocale_t locale = AlarmMessageLocale_t::EN);
 
     void setLastAlarmRequestSuccess(LastCommandSuccess status);
     LastCommandSuccess getLastAlarmRequestSuccess();
@@ -44,6 +52,7 @@ public:
 
 private:
     static int getTimezoneOffset();
+    String getLocaleMessage(const AlarmMessage_t *msg, AlarmMessageLocale_t locale);
 
     uint8_t _payloadAlarmLog[ALARM_LOG_PAYLOAD_SIZE];
     uint8_t _alarmLogLength = 0;
